src/requirements.txt


##These dependencies are brought in by django CMS, but if you want to lock-in their version, specify them
#Django==1.5.1
#django-classy-tags==0.4
#South==0.8.1
#html5lib==1.0b1
#django-mptt==0.5.2
#django-treeadmin
#django-sekizai==0.7
#six==1.3.0
#djangocms-admin-style==0.1.2

#Django CMS repo
git+git://github.com/divio/django-cms.git#egg=cms
#git+git://github.com/divio/django-shop.git#egg=shop

##First install from local
#django-cms
#django-shop
#cmsplugin-shop
#django-shop-configurableproduct
#django-shop-discounts


##Optional, recommended packages
#Pillow
#gunicorn
jsonfield
django-reversion
djangocms_text_ckeditor

django-templateaddons

#django-cms==3.0
django-shop
django-shop-categories
django-shop-simplevariations
django-shop-paypal
django-shop-discounts ##Pending update cloned in local
django-shop-ajax ##adds ajax add to cart
django-shop-simplenotifications ##email notifications
#django-shop-l10n ##Needed for area-tax
#django-shop-area-tax ##Taxes per address
django-appconf
#cmsplugin-shop ##local modified to work from https://github.com/ZibMedia/cmsplugin-shop
#git+git://github.com/airtonix/django-shop-configurableproduct.git#egg=django_shop_configurableproduct
#https://github.com/airtonix/django-shop-configurableproduct ##configurable product if needed
django-shop-configurableproduct
#cmsplugin-configurableproduct


git+git://github.com/philomat/django-form-designer.git#egg=django-form-designer
#cmsplugin-filer==0.9.5
cmsplugin-filer
#git+git://github.com/pbs/cmsplugin-filer.git#egg=cmsplugin_filer
#git+git://github.com/Troyhy/cmsplugin-filer-gallery.git#egg=cmsplugin_filer_gallery
#git+git://github.com/centralniak/cmsplugin_gallery#egg=cmsplugin_gallery
#git+git://github.com/hovel/imagestore#egg=imagestore

#cmsplugin-comments
cmsplugin-embedded-menu
cmsplugin-htmlsitemap
cmsplugin-twitter

#git+git://github.com/netrocks/django-cms-social-networks#egg=cms_social_facebook

django-blog-zinnia
cmsplugin_zinnia



#For development
django-debug-toolbar
django_extensions
Werkzeug

